feat: [CDP-19640]: Refactor K8sStepHelper and NativeHelmStepHelper (#29568)
* feat: [CDP-19640]: Refactor methods/fields used for SCM optimised file fetch

* feat: [CDP-19640]: Refactor methods/fields used for TimeOut and ParameterFieldBoolean

* feat: [CDP-19640]: Remove unused imports

* feat: [CDP-19640]: changes based on review comments
