+++ 125-cd-nextgen/src/main/java/io/harness/cdng/helm/HelmDeployStep.java
@@ -196,6 +196,8 @@ public class HelmDeployStep extends TaskChainExecutableWithRollbackAndRbac imple
             .namespace(nativeHelmStepHelper.getK8sInfraDelegateConfig(infrastructure, ambiance).getNamespace())
             .k8SteadyStateCheckEnabled(cdFeatureFlagHelper.isEnabled(
                 AmbianceUtils.getAccountId(ambiance), FeatureName.HELM_STEADY_STATE_CHECK_1_16))
+            .useLatestKubectlVersion(
+                cdFeatureFlagHelper.isEnabled(AmbianceUtils.getAccountId(ambiance), FeatureName.NEW_KUBECTL_VERSION))
             .shouldOpenFetchFilesLogStream(true)
             .build();
