+++ 125-cd-nextgen/src/main/java/io/harness/cdng/k8s/K8sApplyStep.java
@@ -5,6 +5,8 @@ import static io.harness.data.structure.EmptyPredicate.isEmpty;
 
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
+import io.harness.cdng.ParameterFieldBooleanValueHelper;
+import io.harness.cdng.TimeOutHelper;
 import io.harness.cdng.infra.beans.InfrastructureOutcome;
 import io.harness.cdng.k8s.K8sApplyBaseStepInfo.K8sApplyBaseStepInfoKeys;
 import io.harness.cdng.k8s.beans.GitFetchResponsePassThroughData;
@@ -101,10 +103,10 @@ public class K8sApplyStep extends TaskChainExecutableWithRollbackAndRbac impleme
     InfrastructureOutcome infrastructure = executionPassThroughData.getInfrastructure();
     String releaseName = k8sStepHelper.getReleaseName(ambiance, infrastructure);
     K8sApplyStepParameters k8sApplyStepParameters = (K8sApplyStepParameters) stepElementParameters.getSpec();
-    boolean skipDryRun = K8sStepHelper.getParameterFieldBooleanValue(
+    boolean skipDryRun = ParameterFieldBooleanValueHelper.getParameterFieldBooleanValue(
         k8sApplyStepParameters.getSkipDryRun(), K8sApplyBaseStepInfoKeys.skipDryRun, stepElementParameters);
     boolean skipSteadyStateCheck =
-        K8sStepHelper.getParameterFieldBooleanValue(k8sApplyStepParameters.getSkipSteadyStateCheck(),
+            ParameterFieldBooleanValueHelper.getParameterFieldBooleanValue(k8sApplyStepParameters.getSkipSteadyStateCheck(),
             K8sApplyBaseStepInfoKeys.skipSteadyStateCheck, stepElementParameters);
 
     final String accountId = AmbianceUtils.getAccountId(ambiance);
@@ -114,7 +116,7 @@ public class K8sApplyStep extends TaskChainExecutableWithRollbackAndRbac impleme
             .releaseName(releaseName)
             .commandName(K8S_APPLY_COMMAND_NAME)
             .taskType(K8sTaskType.APPLY)
-            .timeoutIntervalInMin(K8sStepHelper.getTimeoutInMin(stepElementParameters))
+            .timeoutIntervalInMin(TimeOutHelper.getTimeoutInMin(stepElementParameters))
             .valuesYamlList(k8sStepHelper.renderValues(k8sManifestOutcome, ambiance, manifestOverrideContents))
             .k8sInfraDelegateConfig(k8sStepHelper.getK8sInfraDelegateConfig(infrastructure, ambiance))
             .kustomizePatchesList(k8sStepHelper.renderPatches(k8sManifestOutcome, ambiance, manifestOverrideContents))
