+++ 125-cd-nextgen/src/main/java/io/harness/cdng/k8s/K8sApplyStepParameters.java
@@ -1,10 +1,10 @@
 package io.harness.cdng.k8s;
 
 import static io.harness.annotations.dev.HarnessTeam.CDP;
-import static io.harness.cdng.k8s.K8sStepHelper.getParameterFieldBooleanValue;
 
 import io.harness.annotation.RecasterAlias;
 import io.harness.annotations.dev.OwnedBy;
+import io.harness.cdng.ParameterFieldBooleanValueHelper;
 import io.harness.executions.steps.ExecutionNodeType;
 import io.harness.k8s.K8sCommandUnitConstants;
 import io.harness.plancreator.steps.TaskSelectorYaml;
@@ -38,7 +38,7 @@ public class K8sApplyStepParameters extends K8sApplyBaseStepInfo implements K8sS
   @JsonIgnore
   public List<String> getCommandUnits() {
     if (!ParameterField.isNull(skipSteadyStateCheck)
-        && getParameterFieldBooleanValue(skipSteadyStateCheck, K8sApplyBaseStepInfoKeys.skipSteadyStateCheck,
+        && ParameterFieldBooleanValueHelper.getParameterFieldBooleanValue(skipSteadyStateCheck, K8sApplyBaseStepInfoKeys.skipSteadyStateCheck,
             String.format("%s step", ExecutionNodeType.K8S_APPLY.getYamlType()))) {
       return Arrays.asList(K8sCommandUnitConstants.FetchFiles, K8sCommandUnitConstants.Init,
           K8sCommandUnitConstants.Prepare, K8sCommandUnitConstants.Apply, K8sCommandUnitConstants.WrapUp);
