+++ 125-cd-nextgen/src/main/java/io/harness/cdng/k8s/K8sBlueGreenStep.java
@@ -2,6 +2,8 @@ package io.harness.cdng.k8s;
 
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
+import io.harness.cdng.ParameterFieldBooleanValueHelper;
+import io.harness.cdng.TimeOutHelper;
 import io.harness.cdng.infra.beans.InfrastructureOutcome;
 import io.harness.cdng.instance.info.InstanceInfoService;
 import io.harness.cdng.k8s.K8sBlueGreenBaseStepInfo.K8sBlueGreenBaseStepInfoKeys;
@@ -89,7 +91,7 @@ public class K8sBlueGreenStep extends TaskChainExecutableWithRollbackAndRbac imp
     String releaseName = k8sStepHelper.getReleaseName(ambiance, infrastructure);
     K8sBlueGreenStepParameters k8sBlueGreenStepParameters =
         (K8sBlueGreenStepParameters) stepElementParameters.getSpec();
-    boolean skipDryRun = K8sStepHelper.getParameterFieldBooleanValue(
+    boolean skipDryRun = ParameterFieldBooleanValueHelper.getParameterFieldBooleanValue(
         k8sBlueGreenStepParameters.getSkipDryRun(), K8sBlueGreenBaseStepInfoKeys.skipDryRun, stepElementParameters);
     List<String> manifestFilesContents =
         k8sStepHelper.renderValues(k8sManifestOutcome, ambiance, manifestOverrideContents);
@@ -103,7 +105,7 @@ public class K8sBlueGreenStep extends TaskChainExecutableWithRollbackAndRbac imp
             .releaseName(releaseName)
             .commandName(K8S_BLUE_GREEN_DEPLOY_COMMAND_NAME)
             .taskType(K8sTaskType.BLUE_GREEN_DEPLOY)
-            .timeoutIntervalInMin(K8sStepHelper.getTimeoutInMin(stepElementParameters))
+            .timeoutIntervalInMin(TimeOutHelper.getTimeoutInMin(stepElementParameters))
             .valuesYamlList(!isOpenshiftTemplate ? manifestFilesContents : Collections.emptyList())
             .openshiftParamList(isOpenshiftTemplate ? manifestFilesContents : Collections.emptyList())
             .kustomizePatchesList(k8sStepHelper.renderPatches(k8sManifestOutcome, ambiance, manifestOverrideContents))
