+++ 125-cd-nextgen/src/main/java/io/harness/cdng/k8s/K8sRollingStep.java
@@ -2,6 +2,8 @@ package io.harness.cdng.k8s;
 
 import io.harness.annotations.dev.HarnessTeam;
 import io.harness.annotations.dev.OwnedBy;
+import io.harness.cdng.ParameterFieldBooleanValueHelper;
+import io.harness.cdng.TimeOutHelper;
 import io.harness.cdng.infra.beans.InfrastructureOutcome;
 import io.harness.cdng.instance.info.InstanceInfoService;
 import io.harness.cdng.k8s.K8sRollingBaseStepInfo.K8sRollingBaseStepInfoKeys;
@@ -85,7 +87,7 @@ public class K8sRollingStep extends TaskChainExecutableWithRollbackAndRbac imple
     InfrastructureOutcome infrastructure = executionPassThroughData.getInfrastructure();
     String releaseName = k8sStepHelper.getReleaseName(ambiance, infrastructure);
     K8sRollingStepParameters k8sRollingStepParameters = (K8sRollingStepParameters) stepElementParameters.getSpec();
-    boolean skipDryRun = K8sStepHelper.getParameterFieldBooleanValue(
+    boolean skipDryRun = ParameterFieldBooleanValueHelper.getParameterFieldBooleanValue(
         k8sRollingStepParameters.getSkipDryRun(), K8sRollingBaseStepInfoKeys.skipDryRun, stepElementParameters);
     List<String> manifestFilesContents =
         k8sStepHelper.renderValues(k8sManifestOutcome, ambiance, manifestOverrideContents);
@@ -109,7 +111,7 @@ public class K8sRollingStep extends TaskChainExecutableWithRollbackAndRbac imple
             .commandName(K8S_ROLLING_DEPLOY_COMMAND_NAME)
             .taskType(K8sTaskType.DEPLOYMENT_ROLLING)
             .localOverrideFeatureFlag(false)
-            .timeoutIntervalInMin(K8sStepHelper.getTimeoutInMin(stepElementParameters))
+            .timeoutIntervalInMin(TimeOutHelper.getTimeoutInMin(stepElementParameters))
             .valuesYamlList(!isOpenshiftTemplate ? manifestFilesContents : Collections.emptyList())
             .openshiftParamList(isOpenshiftTemplate ? manifestFilesContents : Collections.emptyList())
             .kustomizePatchesList(k8sStepHelper.renderPatches(k8sManifestOutcome, ambiance, manifestOverrideContents))
