+++ 125-cd-nextgen/src/main/java/io/harness/cdng/k8s/K8sScaleStepParameter.java
@@ -1,10 +1,10 @@
 package io.harness.cdng.k8s;
 
 import static io.harness.annotations.dev.HarnessTeam.CDP;
-import static io.harness.cdng.k8s.K8sStepHelper.getParameterFieldBooleanValue;
 
 import io.harness.annotation.RecasterAlias;
 import io.harness.annotations.dev.OwnedBy;
+import io.harness.cdng.ParameterFieldBooleanValueHelper;
 import io.harness.executions.steps.ExecutionNodeType;
 import io.harness.k8s.K8sCommandUnitConstants;
 import io.harness.plancreator.steps.TaskSelectorYaml;
@@ -39,7 +39,7 @@ public class K8sScaleStepParameter extends K8sScaleBaseStepInfo implements K8sSp
   @JsonIgnore
   public List<String> getCommandUnits() {
     if (!ParameterField.isNull(skipSteadyStateCheck)
-        && getParameterFieldBooleanValue(skipSteadyStateCheck, K8sScaleBaseStepInfoKeys.skipSteadyStateCheck,
+        && ParameterFieldBooleanValueHelper.getParameterFieldBooleanValue(skipSteadyStateCheck, K8sScaleBaseStepInfoKeys.skipSteadyStateCheck,
             String.format("%s step", ExecutionNodeType.K8S_SCALE.getYamlType()))) {
       return Arrays.asList(K8sCommandUnitConstants.Init, K8sCommandUnitConstants.Scale, K8sCommandUnitConstants.WrapUp);
     } else {
