+++ 125-cd-nextgen/src/test/java/io/harness/cdng/k8s/K8sStepHelperTest.java
@@ -3,7 +3,6 @@ package io.harness.cdng.k8s;
 import static io.harness.annotations.dev.HarnessTeam.CDP;
 import static io.harness.beans.EnvironmentType.NON_PROD;
 import static io.harness.beans.EnvironmentType.PROD;
-import static io.harness.beans.FeatureName.OPTIMIZED_GIT_FETCH_FILES;
 import static io.harness.cdng.k8s.K8sStepHelper.K8S_SUPPORTED_MANIFEST_TYPES;
 import static io.harness.cdng.k8s.K8sStepHelper.MISSING_INFRASTRUCTURE_ERROR;
 import static io.harness.cdng.k8s.K8sStepHelper.RELEASE_NAME;
@@ -20,7 +19,6 @@ import static io.harness.rule.OwnerRule.ACASIAN;
 import static io.harness.rule.OwnerRule.ACHYUTH;
 import static io.harness.rule.OwnerRule.ANSHUL;
 import static io.harness.rule.OwnerRule.NAMAN_TALAYCHA;
-import static io.harness.rule.OwnerRule.TMACARI;
 import static io.harness.rule.OwnerRule.VAIBHAV_SI;
 import static io.harness.rule.OwnerRule.VIKAS_S;
 
@@ -45,6 +43,7 @@ import io.harness.CategoryTest;
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.beans.FeatureName;
 import io.harness.category.element.UnitTests;
+import io.harness.cdng.ParameterFieldBooleanValueHelper;
 import io.harness.cdng.ReleaseNameHelper;
 import io.harness.cdng.common.beans.SetupAbstractionKeys;
 import io.harness.cdng.featureFlag.CDFeatureFlagHelper;
@@ -94,14 +93,7 @@ import io.harness.delegate.beans.connector.k8Connector.KubernetesCredentialType;
 import io.harness.delegate.beans.connector.scm.GitAuthType;
 import io.harness.delegate.beans.connector.scm.GitConnectionType;
 import io.harness.delegate.beans.connector.scm.genericgitconnector.GitConfigDTO;
-import io.harness.delegate.beans.connector.scm.github.GithubApiAccessDTO;
-import io.harness.delegate.beans.connector.scm.github.GithubAuthenticationDTO;
 import io.harness.delegate.beans.connector.scm.github.GithubConnectorDTO;
-import io.harness.delegate.beans.connector.scm.github.GithubHttpAuthenticationType;
-import io.harness.delegate.beans.connector.scm.github.GithubHttpCredentialsDTO;
-import io.harness.delegate.beans.connector.scm.github.GithubTokenSpecDTO;
-import io.harness.delegate.beans.connector.scm.github.GithubUsernamePasswordDTO;
-import io.harness.delegate.beans.connector.scm.github.GithubUsernameTokenDTO;
 import io.harness.delegate.beans.executioncapability.GitConnectionNGCapability;
 import io.harness.delegate.beans.logstreaming.UnitProgressData;
 import io.harness.delegate.beans.storeconfig.FetchType;
@@ -125,7 +117,6 @@ import io.harness.delegate.task.k8s.KustomizeManifestDelegateConfig;
 import io.harness.delegate.task.k8s.ManifestDelegateConfig;
 import io.harness.delegate.task.k8s.ManifestType;
 import io.harness.delegate.task.k8s.OpenshiftManifestDelegateConfig;
-import io.harness.encryption.SecretRefData;
 import io.harness.eventsframework.schemas.entity.EntityDetailProtoDTO;
 import io.harness.exception.ExceptionUtils;
 import io.harness.exception.GeneralException;
@@ -163,7 +154,6 @@ import io.harness.pms.sdk.core.steps.io.StepResponse;
 import io.harness.pms.yaml.ParameterField;
 import io.harness.rule.Owner;
 import io.harness.secretmanagerclient.services.api.SecretManagerClientService;
-import io.harness.security.encryption.EncryptedDataDetail;
 import io.harness.serializer.KryoSerializer;
 import io.harness.steps.EntityReferenceExtractorUtils;
 import io.harness.steps.StepHelper;
@@ -603,214 +593,6 @@ public class K8sStepHelperTest extends CategoryTest {
   }
 
   @Test
-  @Owner(developers = TMACARI)
-  @Category(UnitTests.class)
-  public void shouldGetGitStoreDelegateConfigFFEnabledGithubUsernameTokenAuth() {
-    List<String> paths = asList("path/to");
-    GitStoreConfig gitStoreConfig = GithubStore.builder()
-                                        .repoName(ParameterField.createValueField("parent-repo/module"))
-                                        .paths(ParameterField.createValueField(paths))
-                                        .build();
-    ConnectorInfoDTO connectorInfoDTO = ConnectorInfoDTO.builder().build();
-    SSHKeySpecDTO sshKeySpecDTO = SSHKeySpecDTO.builder().build();
-    List<EncryptedDataDetail> apiEncryptedDataDetails = new ArrayList<>();
-    GithubConnectorDTO githubConnectorDTO =
-        GithubConnectorDTO.builder()
-            .connectionType(GitConnectionType.ACCOUNT)
-            .url("http://localhost")
-            .authentication(GithubAuthenticationDTO.builder()
-                                .authType(GitAuthType.HTTP)
-                                .credentials(GithubHttpCredentialsDTO.builder()
-                                                 .type(GithubHttpAuthenticationType.USERNAME_AND_TOKEN)
-                                                 .httpCredentialsSpec(GithubUsernameTokenDTO.builder()
-                                                                          .username("usermane")
-                                                                          .tokenRef(SecretRefData.builder().build())
-                                                                          .build())
-                                                 .build())
-                                .build())
-            .build();
-    connectorInfoDTO.setConnectorConfig(githubConnectorDTO);
-
-    doReturn(true).when(cdFeatureFlagHelper).isEnabled(any(), eq(OPTIMIZED_GIT_FETCH_FILES));
-    doReturn(sshKeySpecDTO).when(gitConfigAuthenticationInfoHelper).getSSHKey(any(), any(), any(), any());
-    doReturn(apiEncryptedDataDetails).when(secretManagerClientService).getEncryptionDetails(any(), any());
-
-    GitStoreDelegateConfig gitStoreDelegateConfig = k8sStepHelper.getGitStoreDelegateConfig(
-        gitStoreConfig, connectorInfoDTO, K8sManifestOutcome.builder().build(), paths, ambiance);
-
-    assertThat(gitStoreDelegateConfig).isNotNull();
-    assertThat(gitStoreDelegateConfig.isOptimizedFilesFetch()).isTrue();
-    assertThat(gitStoreDelegateConfig.getGitConfigDTO()).isInstanceOf(GithubConnectorDTO.class);
-    assertThat(gitStoreDelegateConfig.getApiAuthEncryptedDataDetails()).isEqualTo(apiEncryptedDataDetails);
-    GithubConnectorDTO convertedGithubConnectorDTO = (GithubConnectorDTO) gitStoreDelegateConfig.getGitConfigDTO();
-    assertThat(convertedGithubConnectorDTO.getUrl()).isEqualTo("http://localhost/parent-repo/module");
-    assertThat(convertedGithubConnectorDTO.getConnectionType()).isEqualTo(GitConnectionType.REPO);
-    assertThat(convertedGithubConnectorDTO.getApiAccess()).isNotNull();
-  }
-
-  @Test
-  @Owner(developers = TMACARI)
-  @Category(UnitTests.class)
-  public void shouldGetGitStoreDelegateConfigFFEnabledGithubApiAccess() {
-    List<String> paths = asList("path/to");
-    GitStoreConfig gitStoreConfig = GithubStore.builder()
-                                        .repoName(ParameterField.createValueField("parent-repo/module"))
-                                        .paths(ParameterField.createValueField(paths))
-                                        .build();
-    ConnectorInfoDTO connectorInfoDTO = ConnectorInfoDTO.builder().build();
-    SSHKeySpecDTO sshKeySpecDTO = SSHKeySpecDTO.builder().build();
-    List<EncryptedDataDetail> apiEncryptedDataDetails = new ArrayList<>();
-    GithubConnectorDTO githubConnectorDTO =
-        GithubConnectorDTO.builder()
-            .connectionType(GitConnectionType.ACCOUNT)
-            .url("http://localhost")
-            .apiAccess(GithubApiAccessDTO.builder().spec(GithubTokenSpecDTO.builder().build()).build())
-            .authentication(GithubAuthenticationDTO.builder()
-                                .authType(GitAuthType.HTTP)
-                                .credentials(GithubHttpCredentialsDTO.builder()
-                                                 .type(GithubHttpAuthenticationType.USERNAME_AND_PASSWORD)
-                                                 .httpCredentialsSpec(GithubUsernamePasswordDTO.builder()
-                                                                          .username("usermane")
-                                                                          .passwordRef(SecretRefData.builder().build())
-                                                                          .build())
-                                                 .build())
-                                .build())
-            .build();
-    connectorInfoDTO.setConnectorConfig(githubConnectorDTO);
-
-    doReturn(true).when(cdFeatureFlagHelper).isEnabled(any(), eq(OPTIMIZED_GIT_FETCH_FILES));
-    doReturn(sshKeySpecDTO).when(gitConfigAuthenticationInfoHelper).getSSHKey(any(), any(), any(), any());
-    doReturn(apiEncryptedDataDetails).when(secretManagerClientService).getEncryptionDetails(any(), any());
-
-    GitStoreDelegateConfig gitStoreDelegateConfig = k8sStepHelper.getGitStoreDelegateConfig(
-        gitStoreConfig, connectorInfoDTO, K8sManifestOutcome.builder().build(), paths, ambiance);
-
-    assertThat(gitStoreDelegateConfig).isNotNull();
-    assertThat(gitStoreDelegateConfig.isOptimizedFilesFetch()).isTrue();
-    assertThat(gitStoreDelegateConfig.getGitConfigDTO()).isInstanceOf(GithubConnectorDTO.class);
-    assertThat(gitStoreDelegateConfig.getApiAuthEncryptedDataDetails()).isEqualTo(apiEncryptedDataDetails);
-    GithubConnectorDTO convertedGithubConnectorDTO = (GithubConnectorDTO) gitStoreDelegateConfig.getGitConfigDTO();
-    assertThat(convertedGithubConnectorDTO.getUrl()).isEqualTo("http://localhost/parent-repo/module");
-    assertThat(convertedGithubConnectorDTO.getConnectionType()).isEqualTo(GitConnectionType.REPO);
-    assertThat(convertedGithubConnectorDTO.getApiAccess()).isNotNull();
-  }
-
-  @Test
-  @Owner(developers = NAMAN_TALAYCHA)
-  @Category(UnitTests.class)
-  public void shouldGetGitStoreDelegateConfigOptimizedFilesFetchWithKustomize() {
-    List<String> paths = asList("path/to");
-    GitStoreConfig gitStoreConfig = GithubStore.builder()
-                                        .repoName(ParameterField.createValueField("parent-repo/module"))
-                                        .paths(ParameterField.createValueField(paths))
-                                        .build();
-    ConnectorInfoDTO connectorInfoDTO = ConnectorInfoDTO.builder().build();
-    SSHKeySpecDTO sshKeySpecDTO = SSHKeySpecDTO.builder().build();
-    List<EncryptedDataDetail> apiEncryptedDataDetails = new ArrayList<>();
-    GithubConnectorDTO githubConnectorDTO =
-        GithubConnectorDTO.builder()
-            .connectionType(GitConnectionType.ACCOUNT)
-            .url("http://localhost")
-            .apiAccess(GithubApiAccessDTO.builder().spec(GithubTokenSpecDTO.builder().build()).build())
-            .authentication(GithubAuthenticationDTO.builder()
-                                .authType(GitAuthType.HTTP)
-                                .credentials(GithubHttpCredentialsDTO.builder()
-                                                 .type(GithubHttpAuthenticationType.USERNAME_AND_PASSWORD)
-                                                 .httpCredentialsSpec(GithubUsernamePasswordDTO.builder()
-                                                                          .username("usermane")
-                                                                          .passwordRef(SecretRefData.builder().build())
-                                                                          .build())
-                                                 .build())
-                                .build())
-            .build();
-    connectorInfoDTO.setConnectorConfig(githubConnectorDTO);
-
-    doReturn(true).when(cdFeatureFlagHelper).isEnabled(any(), eq(OPTIMIZED_GIT_FETCH_FILES));
-    doReturn(sshKeySpecDTO).when(gitConfigAuthenticationInfoHelper).getSSHKey(any(), any(), any(), any());
-    doReturn(apiEncryptedDataDetails).when(secretManagerClientService).getEncryptionDetails(any(), any());
-
-    GitStoreDelegateConfig gitStoreDelegateConfig = k8sStepHelper.getGitStoreDelegateConfig(
-        gitStoreConfig, connectorInfoDTO, KustomizeManifestOutcome.builder().build(), paths, ambiance);
-
-    assertThat(gitStoreDelegateConfig).isNotNull();
-    assertThat(gitStoreDelegateConfig.isOptimizedFilesFetch()).isFalse();
-    GitConfigDTO gitConfigDTO = (GitConfigDTO) gitStoreDelegateConfig.getGitConfigDTO();
-    assertThat(gitConfigDTO.getUrl()).isEqualTo("http://localhost/parent-repo/module");
-    assertThat(gitConfigDTO.getGitConnectionType()).isEqualTo(GitConnectionType.REPO);
-  }
-
-  @Test
-  @Owner(developers = TMACARI)
-  @Category(UnitTests.class)
-  public void shouldGetGitStoreDelegateConfigFFEnabledNotGithubOrGitlab() {
-    List<String> paths = asList("path/to");
-    GitStoreConfig gitStoreConfig = GithubStore.builder()
-                                        .repoName(ParameterField.createValueField("parent-repo/module"))
-                                        .paths(ParameterField.createValueField(paths))
-                                        .build();
-    ConnectorInfoDTO connectorInfoDTO = ConnectorInfoDTO.builder().build();
-    SSHKeySpecDTO sshKeySpecDTO = SSHKeySpecDTO.builder().build();
-    GitConfigDTO gitConfigDTO =
-        GitConfigDTO.builder().gitConnectionType(GitConnectionType.ACCOUNT).url("http://localhost").build();
-    connectorInfoDTO.setConnectorConfig(gitConfigDTO);
-
-    doReturn(true).when(cdFeatureFlagHelper).isEnabled(any(), eq(OPTIMIZED_GIT_FETCH_FILES));
-    doReturn(sshKeySpecDTO).when(gitConfigAuthenticationInfoHelper).getSSHKey(any(), any(), any(), any());
-
-    GitStoreDelegateConfig gitStoreDelegateConfig = k8sStepHelper.getGitStoreDelegateConfig(
-        gitStoreConfig, connectorInfoDTO, K8sManifestOutcome.builder().build(), paths, ambiance);
-
-    assertThat(gitStoreDelegateConfig).isNotNull();
-    assertThat(gitStoreDelegateConfig.getGitConfigDTO()).isInstanceOf(GitConfigDTO.class);
-    assertThat(gitStoreDelegateConfig.isOptimizedFilesFetch()).isFalse();
-    GitConfigDTO convertedConfig = (GitConfigDTO) gitStoreDelegateConfig.getGitConfigDTO();
-    assertThat(convertedConfig.getUrl()).isEqualTo("http://localhost/parent-repo/module");
-    assertThat(convertedConfig.getGitConnectionType()).isEqualTo(GitConnectionType.REPO);
-  }
-
-  @Test
-  @Owner(developers = TMACARI)
-  @Category(UnitTests.class)
-  public void shouldGetGitStoreDelegateConfigFFEnabledGithubUsernamePasswordAuth() {
-    List<String> paths = asList("path/to");
-    GitStoreConfig gitStoreConfig = GithubStore.builder()
-                                        .repoName(ParameterField.createValueField("parent-repo/module"))
-                                        .paths(ParameterField.createValueField(paths))
-                                        .build();
-    ConnectorInfoDTO connectorInfoDTO = ConnectorInfoDTO.builder().build();
-    SSHKeySpecDTO sshKeySpecDTO = SSHKeySpecDTO.builder().build();
-    GithubConnectorDTO githubConnectorDTO =
-        GithubConnectorDTO.builder()
-            .connectionType(GitConnectionType.ACCOUNT)
-            .url("http://localhost")
-            .authentication(GithubAuthenticationDTO.builder()
-                                .authType(GitAuthType.HTTP)
-                                .credentials(GithubHttpCredentialsDTO.builder()
-                                                 .type(GithubHttpAuthenticationType.USERNAME_AND_PASSWORD)
-                                                 .httpCredentialsSpec(GithubUsernamePasswordDTO.builder()
-                                                                          .username("usermane")
-                                                                          .passwordRef(SecretRefData.builder().build())
-                                                                          .build())
-                                                 .build())
-                                .build())
-            .build();
-    connectorInfoDTO.setConnectorConfig(githubConnectorDTO);
-
-    doReturn(true).when(cdFeatureFlagHelper).isEnabled(any(), eq(OPTIMIZED_GIT_FETCH_FILES));
-    doReturn(sshKeySpecDTO).when(gitConfigAuthenticationInfoHelper).getSSHKey(any(), any(), any(), any());
-
-    GitStoreDelegateConfig gitStoreDelegateConfig = k8sStepHelper.getGitStoreDelegateConfig(
-        gitStoreConfig, connectorInfoDTO, K8sManifestOutcome.builder().build(), paths, ambiance);
-
-    assertThat(gitStoreDelegateConfig).isNotNull();
-    assertThat(gitStoreDelegateConfig.getGitConfigDTO()).isInstanceOf(GitConfigDTO.class);
-    assertThat(gitStoreDelegateConfig.isOptimizedFilesFetch()).isFalse();
-    GitConfigDTO convertedConfig = (GitConfigDTO) gitStoreDelegateConfig.getGitConfigDTO();
-    assertThat(convertedConfig.getUrl()).isEqualTo("http://localhost/parent-repo/module");
-    assertThat(convertedConfig.getGitConnectionType()).isEqualTo(GitConnectionType.REPO);
-  }
-
-  @Test
   @Owner(developers = ACASIAN)
   @Category(UnitTests.class)
   public void shouldNotConvertGitRepoWithRepoName() {
@@ -1102,35 +884,6 @@ public class K8sStepHelperTest extends CategoryTest {
     assertThat(gcsHelmStoreDelegateConfig.getBucketName()).isEqualTo(bucketName);
     assertThat(gcsHelmStoreDelegateConfig.getFolderPath()).isEqualTo(folderPath);
     assertThat(gcsHelmStoreDelegateConfig.getGcpConnector()).isEqualTo(gcpConnectorDTO);
-  }
-
-  @Test
-  @Owner(developers = ABOSII)
-  @Category(UnitTests.class)
-  public void testGetTimeoutValue() {
-    StepElementParameters definedValue =
-        StepElementParameters.builder().timeout(ParameterField.createValueField("15m")).build();
-    StepElementParameters nullValue = StepElementParameters.builder().timeout(ParameterField.ofNull()).build();
-    assertThat(K8sStepHelper.getTimeoutValue(definedValue)).isEqualTo("15m");
-    assertThat(K8sStepHelper.getTimeoutValue(nullValue)).isEqualTo("10m");
-  }
-
-  @Test
-  @Owner(developers = ABOSII)
-  @Category(UnitTests.class)
-  public void testGetTimeoutInMin() {
-    StepElementParameters value =
-        StepElementParameters.builder().timeout(ParameterField.createValueField("15m")).build();
-    assertThat(K8sStepHelper.getTimeoutInMin(value)).isEqualTo(15);
-  }
-
-  @Test
-  @Owner(developers = ABOSII)
-  @Category(UnitTests.class)
-  public void testGetTimeoutInMillis() {
-    StepElementParameters value =
-        StepElementParameters.builder().timeout(ParameterField.createValueField("15m")).build();
-    assertThat(K8sStepHelper.getTimeoutInMillis(value)).isEqualTo(900000);
   }
 
   @Test
@@ -1643,15 +1396,15 @@ public class K8sStepHelperTest extends CategoryTest {
   @Owner(developers = ABOSII)
   @Category(UnitTests.class)
   public void testGetParameterFieldBooleanValue() {
-    assertThat(K8sStepHelper.getParameterFieldBooleanValue(
+    assertThat(ParameterFieldBooleanValueHelper.getParameterFieldBooleanValue(
                    ParameterField.createValueField("true"), "testField", StepElementParameters.builder().build()))
         .isTrue();
-    assertThat(K8sStepHelper.getParameterFieldBooleanValue(
+    assertThat(ParameterFieldBooleanValueHelper.getParameterFieldBooleanValue(
                    ParameterField.createValueField("false"), "testField", StepElementParameters.builder().build()))
         .isFalse();
 
     assertThatThrownBy(()
-                           -> K8sStepHelper.getParameterFieldBooleanValue(ParameterField.createValueField("absad"),
+                           -> ParameterFieldBooleanValueHelper.getParameterFieldBooleanValue(ParameterField.createValueField("absad"),
                                "testField", StepElementParameters.builder().identifier("test").type("Test").build()))
         .hasMessageContaining("for field testField in Test step with identifier: test");
   }
